pyproject.toml
[project]
name = "jax-gemma-1b"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jax>=0.8.2",
    "google-cloud-storage>=2.18.0",
    "ml-dtypes>=0.5.4",
    "numpy>=2.4.0",
    "optax>=0.2.6",
    "rich>=14.2.0",
    "safetensors>=0.7.0",
    "sentencepiece>=0.2.1",
]

[project.optional-dependencies]
cuda = ["jax[cuda12]>=0.8.2"]
tpu = ["jax[tpu]>=0.8.2"]

[tool.ruff.lint]
ignore = ["E402"]

[[tool.uv.index]]
name = "jax"
url = "https://storage.googleapis.com/jax-releases/libtpu_releases.html"
explicit = false
